refactor(editor): merged MultipleSelect/SingleSelect Editor into 1 class
