+++ client-cli/src/examples/slickgrid/example3.js
@@ -298,7 +298,7 @@ export class Example3 {
 
   onCellChanged(e, args) {
     console.log('onCellChange', args);
-    this.updatedObject = args.item;
+    this.updatedObject = { ...args.item };
   }
 
   onCellClicked(e, args) {
